+++ packages/shared/src/lib/core/nfts/actions/index.ts
@@ -9,6 +9,8 @@ export * from './getPersistedErc721Nfts'
 export * from './interruptNftDownloadAfterTimeout'
 export * from './isNftPersisted'
 export * from './loadNftsForActiveProfile'
+export * from './persistErc721Nft'
+export * from './persistNftWithContractMetadata'
 export * from './setAccountNftsInAllAccountNfts'
 export * from './setNftInAllAccountNftsToUnspendable'
 export * from './stopDownloadingNftMediaFromQueue'
